Dialogue-System
Changes
Renamed Option buttons (1,2,3 to 0,1,2)
Created DialogueTrigger script which triggers dialogue whenever object it is attached to is clicked on
Reworked DialogueManager script completely (made it a singleton)
Added Layers for ClothesPile and Laundry so that Clothes Piles wouldn’t collide with anything other than the Laundry Basket when being dragged (Which was achieved through the LayerCollisionMatrix)
ItemGlow script was modified to include a reference to the object’s starting layer, which is what is used to reset it’s layer
Created unique script that deactivates objects in scene
Created a singleton script that holds the number of loops, good choices, and bad choices the player has made
Modified UI script to include a function for increasing the loop
Cleaned up the TimeChange script, the original code is still there but commented out incase you still want it

Known Bugs
When the close button on the dialogue window is clicked, the next line of dialogue will flash in the text bubble before closing. This is likely due to left click still causing the dialogue to advance.

Any new scripts I have made are in the John’s Scripts folder