Created Bartender dialogue
All options lead to Matt going to the bathroom. As a good option, he can drink water then need to use the toilet so that way the atmosphere doesnt go down. The others would be drinking alcohol until Matt either needs to pee, feels sick and needs to wash up or just gets too drunk that the bartender gives him the life advice and sends him to the bathroom